Atualização de deploy

+++ src/main/java/com/banco/cast/service/BancoService.java
@@ -3,11 +3,13 @@ package com.banco.cast.service;
 import com.banco.cast.model.Conta;
 import com.banco.cast.model.Usuario;
 
+import java.math.BigDecimal;
+
 public interface BancoService {
 
     Conta criarConta(Usuario.UsuarioRequest request);
-    void creditar(String numeroConta, Double valor);
-    void debitar(String numeroConta, Double valor);
-    void transferir(String contaOrigem, String contaDestino, Double valor);
+    void creditar(String numeroConta, BigDecimal valor);
+    void debitar(String numeroConta, BigDecimal valor);
+    void transferir(String contaOrigem, String contaDestino, BigDecimal valor);
     Conta.ExtratoResponse emitirExtrato(String numeroConta);
 }
